# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Swagger CRUD",
        	"description": "Swagger CRUD description",
        	"contact": {
            		"email": "arshkochhar15@gmail.com"
        	}
	},
    	"host": "localhost:5000",
	"basePath": "",
	"tags": [
		{
			"name": "Users",
			"description": "User actions"
		}
	],
	"schemes": ["http", "https"],
	"consumes": ["application/json"],
	"produces": ["application/json"],
    "paths": {
		"/api/v1/inventory": {
			"post": {
				"tags": ["Products"],
				"description": "Create Product",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "item",
						"required": ["name", "category", "price", "quantity", "url"],
						"in": "body",
						"properties": {
							"name": {
								"$ref": "#/definitions/name"
							},
							"category": {
								"$ref": "#/definitions/category"
							},
							"price": {
								"$ref": "#/definitions/price"
							},
                            "quantity": {
								"$ref": "#/definitions/quantity"
							},
                            "url": {
								"$ref": "#/definitions/url"
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Business Exception",
						"schema": {
							"$ref": "#/definitions/business_exception"
						}
					}
				}
			},
			"put": {
				"tags": ["Products"],
				"description": "Update Product",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "item",
						"in": "body",
						"properties": {
							"name": {
								"$ref": "#/definitions/name"
							},
							"category": {
								"$ref": "#/definitions/category"
							},
							"price": {
								"$ref": "#/definitions/price"
							},
                            "quantity": {
								"$ref": "#/definitions/quantity"
							},
                            "url": {
								"$ref": "#/definitions/url"
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Business Exception",
						"schema": {
							"$ref": "#/definitions/business_exception"
						}
					}
				}
			},
            "get": {
				"tags": ["Products"],
				"description": "List Products",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "_id",
						"in": "query",
						"type": "string",
						"description": "Collection ID",
						"example": "5c0728badc70ea3dec00d8e6",
						"defaultValue": ""
					},
					{
						"name": "profile_id",
						"in": "query",
						"type": "number",
						"description": "Profile ID (1 = ADMIN, 2 = CLIENT)",
						"example": 1
					},
					{
						"name": "username",
						"in": "query",
						"type": "string",
						"description": "User logon",
						"example": "LUCIANO"
					},
					{
						"name": "password",
						"in": "query",
						"type": "string",
						"description": "User password",
						"example": "1234"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Business Exception",
						"schema": {
							"$ref": "#/definitions/business_exception"
						}
					}
				}
			},

			"delete": {
				"tags": ["Products"],
				"description": "Delete Product",
				"produces": ["application/json"],
				"parameters": [
					{
						"id": "_id"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Business Exception",
						"schema": {
							"$ref": "#/definitions/business_exception"
						}
					}
				}
			}
		}
	},
	"definitions": {
		"_id": {
			"type": "string",
			"description": "Collection ID",
			"example": "5c0728badc70ea3dec00d8e6"
		},
		"name": {
			"type": "string",
			"description": "Product Name",
			"example": "t-shirt"
		},
		"category": {
			"type": "string",
			"description": "Product Category",
			"example": "clothes"
		},
		"price": {
			"type": "string",
			"description": "Price of Product",
			"example": "22.99"
		},
        "quantity": {
			"type": "string",
			"description": "Quantity of Product",
			"example": "3"
		},
        "url": {
			"type": "string",
			"description": "URL of Product",
			"example": "https://www.amazon.ca/dp/B08X1QWB9W?th=1"
		},
		"error": {
			"type": "object",
			"description": "Error data",
			"properties": {
				"msg": {
					"type": "string",
					"description": "Error message",
					"example": "Product not found"
				}
			}
		},
		"business_exception": {
			"type": "array",
			"description": "Error array",
			"items": {
				"$ref": "#/definitions/error"
			}
		}
	}
}